Moved metric test
