+++ kie-server-parent/kie-server-tests/kie-server-integ-tests-dmn/src/test/java/org/kie/server/integrationtests/dmn/DMNIntegrationTest.java
@@ -17,15 +17,7 @@ package org.kie.server.integrationtests.dmn;
 
 import java.math.BigDecimal;
 import java.util.Map;
-import java.util.concurrent.TimeUnit;
-import javax.ws.rs.client.Client;
-import javax.ws.rs.client.WebTarget;
-import javax.ws.rs.core.MediaType;
-import javax.ws.rs.core.Response;
 
-import org.jboss.resteasy.client.jaxrs.BasicAuthentication;
-import org.jboss.resteasy.client.jaxrs.ResteasyClientBuilder;
-import org.junit.Assert;
 import org.junit.BeforeClass;
 import org.junit.Test;
 import org.kie.api.KieServices;
@@ -35,14 +27,11 @@ import org.kie.dmn.api.core.DMNResult;
 import org.kie.server.api.model.KieServiceResponse.ResponseType;
 import org.kie.server.api.model.ReleaseId;
 import org.kie.server.api.model.ServiceResponse;
-import org.kie.server.integrationtests.config.TestConfig;
 import org.kie.server.integrationtests.shared.KieServerDeployer;
 
 import static org.hamcrest.Matchers.hasEntry;
 import static org.hamcrest.Matchers.is;
-import static org.hamcrest.Matchers.isEmptyOrNullString;
 import static org.hamcrest.Matchers.nullValue;
-import static org.hamcrest.core.IsNot.not;
 import static org.junit.Assert.*;
 
 public class DMNIntegrationTest
@@ -86,59 +75,6 @@ public class DMNIntegrationTest
         
         Map<String, Object> dr0 = (Map<String, Object>) dmnResult.getDecisionResultByName("Math").getResult();
         assertThat( dr0, hasEntry( "Sum", BigDecimal.valueOf( 15 ) ) );
-    }
-
-    private static Client httpClient;
-
-    protected WebTarget newRequest(String uriString) {
-        if (httpClient == null) {
-            httpClient = new ResteasyClientBuilder()
-                    .establishConnectionTimeout(10, TimeUnit.SECONDS)
-                    .socketTimeout(10, TimeUnit.SECONDS)
-                    .build();
-        }
-        WebTarget webTarget = httpClient.target(uriString);
-        webTarget.register(new BasicAuthentication(TestConfig.getUsername(), TestConfig.getPassword()));
-        return webTarget;
-    }
-
-    @Test
-    public void test_evaluateAllWithMetrics() {
-        DMNContext dmnContext = dmnClient.newContext();
-        dmnContext.set( "a", 10 );
-        dmnContext.set( "b", 5 );
-        ServiceResponse<DMNResult> evaluateAll = dmnClient.evaluateAll(CONTAINER_1_ID, dmnContext);
-
-        assertEquals(ResponseType.SUCCESS, evaluateAll.getType());
-
-        DMNResult dmnResult = evaluateAll.getResult();
-
-        Map<String, Object> mathInCtx = (Map<String, Object>) dmnResult.getContext().get( "Math" );
-        assertThat( mathInCtx, hasEntry( "Sum", BigDecimal.valueOf( 15 ) ) );
-
-        Map<String, Object> dr0 = (Map<String, Object>) dmnResult.getDecisionResultByName("Math").getResult();
-        assertThat( dr0, hasEntry( "Sum", BigDecimal.valueOf( 15 ) ) );
-
-
-        Response response = null;
-        try {
-            String uriString = TestConfig.getKieServerHttpUrl() + "/prometheus";
-
-            WebTarget clientRequest = newRequest(uriString);
-            response = clientRequest.request(MediaType.TEXT_PLAIN_TYPE).get();
-
-            Assert.assertEquals(Response.Status.OK.getStatusCode(), response.getStatus());
-
-            String res = response.readEntity(String.class);
-            Assert.assertThat(res, not(isEmptyOrNullString()));
-        } catch (Exception e) {
-            throw new RuntimeException(e);
-        } finally {
-            if (response != null) {
-                response.close();
-            }
-        }
-
     }
     
     // Using explicit namespace and model name
